♻️ small refacto on artist infos
